// honey-paradise-client/src/shared/types/server/getAllConnectionsResponseType.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * Honey Paradise
 * Honey Paradise API
 * OpenAPI spec version: 0.0.2
 */

export type GetAllConnectionsResponseType = typeof GetAllConnectionsResponseType[keyof typeof GetAllConnectionsResponseType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const GetAllConnectionsResponseType = {
  CREDENTIALS: 'CREDENTIALS',
  GOOGLE: 'GOOGLE',
  GITHUB: 'GITHUB',
  VK: 'VK',
  YANDEX: 'YANDEX',
} as const;